作成した合成関数をモジュール化
