+++ section_10/vue3-test/src/composables/useCounter.js
@@ -1,0 +1,19 @@
+import { computed } from 'vue'
+export default function useCounter(item){
+
+  const increment = () => {
+    item.amount++
+  }
+  const decrement = () => {
+    item.amount--
+  }
+  const totalPrice = computed(()=>{
+    return item.price * item.amount
+  })
+
+  return {
+    increment,
+    decrement,
+    totalPrice
+  }
+}
